# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ErrorVo": {
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"messageCode": {
						"type": "string"
					},
					"events": {
						"items": {
							"$ref": "#/components/schemas/EventParamsType"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"message",
					"messageCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"SupportChainId": {
				"enum": [
					137,
					80001,
					56,
					97
				],
				"type": "number"
			},
...
			"ConsoleMember": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"recoveryEmail": {
						"type": "string",
						"nullable": true
					},
					"phoneCertificateYn": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"loginId": {
						"type": "string"
					},
					"platformMemberId": {
						"type": "string"
					},
					"memberId": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"recoveryEmail",
					"phoneCertificateYn",
					"phone",
					"email",
					"nickname",
					"name",
					"loginId",
					"platformMemberId",
					"memberId"
				],
				"type": "object",
				"description": "Model ConsoleMember\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_ConsoleMember_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/ConsoleMember"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateConsoleMemberRequest": {
				"properties": {
					"loginId": {
						"type": "string",
						"example": "test1234",
						"minLength": 8
					},
					"name": {
						"type": "string",
						"description": "이름",
						"example": "홍길동"
					},
					"nickname": {
						"type": "string",
						"description": "닉네임",
						"example": "지트북"
					},
					"email": {
						"type": "string",
						"description": "이메일 주소",
						"example": "test@gmail.com"
					},
					"phone": {
						"type": "string",
						"description": "핸드폰 번호"
					},
					"phoneCertificateYn": {
						"type": "string",
						"description": "모바일 인증 여부"
					},
					"recoveryEmail": {
						"type": "string",
						"description": "복구 이메일"
					}
				},
				"required": [
					"loginId",
					"name",
					"nickname",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_CreateConsoleMemberRequest.loginId-or-platformMemberId_": {
				"$ref": "#/components/schemas/Pick_CreateConsoleMemberRequest.Exclude_keyofCreateConsoleMemberRequest.loginId-or-platformMemberId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UpdateConsoleMemberRequest": {
				"$ref": "#/components/schemas/Omit_CreateConsoleMemberRequest.loginId-or-platformMemberId_"
			},
			"CommonResponse_ConsoleMember-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ConsoleMember"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"NftItemDataType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"imgSrc": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"nftName": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"sales": {
						"type": "boolean"
					},
					"price": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"total": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"rest": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"sales",
					"price",
					"symbol",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_NftItemDataType-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/NftItemDataType"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_NftItemDetailDataType_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/NftItemDetailDataType"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"Workspace": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"workspaceId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"name",
					"workspaceId"
				],
				"type": "object",
				"description": "Model Workspace\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_Workspace_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/Workspace"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateWorkspaceRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "워크스페이스 이름"
					},
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱 id"
					}
				},
				"required": [
					"name",
					"appId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWorkspaceRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "변경할 워크스페이스 이름"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse_FindManyWorkspacesResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FindManyWorkspacesResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_FindManyWorkspaceMembersResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FindManyWorkspaceMembersResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_WorkspaceMember_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/WorkspaceMember"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateWorkspaceMemberRequest": {
				"properties": {
					"workspaceId": {
						"type": "number",
						"format": "double",
						"description": "워크 스페이스 id",
						"example": 1
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "초대할 회원 id",
						"example": 1
					},
					"authorityCode": {
						"type": "string",
						"enum": [
							"MASTER",
							"NORMAL"
						],
						"description": "권한코드. 코드그룹: ACC_AUTH\r\nMASTER: 마스터,\r\nNORMAL: 일반",
						"example": "NORMAL",
						"default": "\"NORMAL\""
					}
				},
				"required": [
					"workspaceId",
					"memberId",
					"authorityCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWorkspaceMemberRequest": {
				"$ref": "#/components/schemas/CreateWorkspaceMemberRequest"
			},
			"DeleteWorkspaceMemberRequest": {
				"properties": {
					"workspaceId": {
						"type": "number",
						"format": "double",
						"description": "워크스페이스 id"
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "회원 id"
					}
				},
				"required": [
					"workspaceId",
					"memberId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InviteWorkspaceMemberRequest": {
				"properties": {
					"workspaceId": {
						"type": "number",
						"format": "double",
						"description": "워크스페이스 id",
						"example": 1
					},
					"email": {
						"type": "string",
						"description": "이메일 주소",
						"example": "test@naver.com"
					},
					"authorityCode": {
						"type": "string",
						"description": "권한코드\r\n코드그룹: ACC_AUTH,\r\nMASTER: 마스터,\r\nNORMAL: 일반",
						"example": "NORMAL"
					}
				},
				"required": [
					"workspaceId",
					"email",
					"authorityCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse_WorkspaceInvitation_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/WorkspaceInvitation"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"UpdateInvitationRequest": {
				"properties": {
					"statusCode": {
						"type": "string",
						"description": "변경 할 상태 값\r\n코드그룹: ACINV_STAT\r\nP: pending,\r\nA: accept,\r\nR: reject,\r\nC: account created",
						"example": "A"
					}
				},
				"required": [
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"devAuth": {
				"type": "http",
				"scheme": "basic"
			},
			"basic": {
				"type": "http",
				"scheme": "basic"
			},
			"token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"cookie": {
				"type": "apiKey",
				"name": "accessToken",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "platpome-console-backend",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/members": {
			"post": {
				"operationId": "CreateConsoleMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_ConsoleMember_"
								}
							}
						}
					}
				},
				"description": "개발자 콘솔 회원 등록 API",
				"tags": [
					"[콘솔 회원] ConsoleMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateConsoleMemberRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindManyConsoleMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_ConsoleMember-Array_"
								}
							}
						}
					}
				},
				"description": "개발자 콘솔 회원 목록 조회 API",
				"tags": [
					"[콘솔 회원] ConsoleMember",
					"[세팅] Setting",
					"[콘솔 회원] ConsoleMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "nickName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/members/{memberId}": {
			"put": {
				"operationId": "UpdateConsoleMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_ConsoleMember_"
								}
							}
						}
					}
				},
				"description": "개발자 콘솔 회원 수정 API",
				"tags": [
					"[콘솔 회원] ConsoleMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "memberId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateConsoleMemberRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindConsoleMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_ConsoleMember_"
								}
							}
						}
					}
				},
				"description": "개발자 콘솔 회원 상세 조회 API",
				"tags": [
					"[콘솔 회원] ConsoleMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "memberId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/chains/{chainId}/erc1155": {
			"get": {
				"operationId": "GetNFTs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_NftItemDataType-Array_"
								}
							}
						}
					}
				},
				"description": "NFT 목록 조회",
				"tags": [
					"[NFT] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "97:BSC testnet, 56:BSC mainnet, 80001:mumbai testnet, 137:polygon mainnet",
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					}
				]
			}
		},
		"/chains/{chainId}/wallets/{walletAddress}/erc1155": {
			"get": {
				"operationId": "GetOwnedNFTs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_NftItemDataType-Array_"
								}
							}
						}
					}
				},
				"description": "NFT 목록 지갑주소로 조회",
				"tags": [
					"[NFT] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "97:BSC testnet, 56:BSC mainnet, 80001:mumbai testnet, 137:polygon mainnet",
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					},
					{
						"description": "지갑주소",
						"in": "path",
						"name": "walletAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chains/{chainId}/erc1155/{tokenId}": {
			"get": {
				"operationId": "GetNFT",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_NftItemDetailDataType_"
								}
							}
						}
					}
				},
				"tags": [
					"[NFT] Common"
				],
				"security": [
					{
						"cookie": []
					},
					{
						"token": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					},
					{
						"in": "path",
						"name": "tokenId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/workspaces": {
			"post": {
				"operationId": "CreateWorkspace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_Workspace_"
								}
							}
						}
					}
				},
				"description": "** 0.1 버전에서는 워크스페이스 생성,수정 기획 없음 **\r\n** app 생성시 default 워크스페이스 생성 **\r\n워크스페이스 생성 API",
				"tags": [
					"[워크 스페이스] Workspace"
				],
				"deprecated": true,
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWorkspaceRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindManyWorkspaces",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindManyWorkspacesResponse-Array_"
								}
							}
						}
					}
				},
				"description": "** 0.1 버전에서는 워크스페이스 생성,수정 기획 없음 **\r\n앱 하위 워크스페이스 목록 조회 API",
				"tags": [
					"[워크 스페이스] Workspace"
				],
				"deprecated": true,
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/workspaces/{id}": {
			"put": {
				"operationId": "UpdateWorkspace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_Workspace_"
								}
							}
						}
					}
				},
				"description": "** 0.1 버전에서는 워크스페이스 생성,수정 기획 없음 **\r\n워크스페이스 수정 API",
				"tags": [
					"[워크 스페이스] Workspace"
				],
				"deprecated": true,
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkspaceRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindWorkspace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_Workspace_"
								}
							}
						}
					}
				},
				"description": "** 0.1 버전에서는 워크스페이스 생성,수정 기획 없음 **\r\n워크스페이스 상세 조회 API",
				"tags": [
					"[워크 스페이스] Workspace"
				],
				"deprecated": true,
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/workspace-members": {
			"get": {
				"operationId": "FindManyWorkspaceMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindManyWorkspaceMembersResponse-Array_"
								}
							}
						}
					}
				},
				"description": "Setting > 팀원 및 권한 > 팀원 목록 조회 API",
				"tags": [
					"[세팅] Setting",
					"[워크 스페이스 팀원] WorkspaceMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateWorkspaceMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_WorkspaceMember_"
								}
							}
						}
					}
				},
				"description": "Setting > 팀원 및 권한 > 워크스페이스 소속 팀원 추가 API",
				"tags": [
					"[세팅] Setting",
					"[워크 스페이스 팀원] WorkspaceMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWorkspaceMemberRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateWorkspaceMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_WorkspaceMember_"
								}
							}
						}
					}
				},
				"description": "워크스페이스 소속 팀원 권한 수정 API",
				"tags": [
					"[워크 스페이스 팀원] WorkspaceMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkspaceMemberRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWorkspaceMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse"
								}
							}
						}
					}
				},
				"description": "워크스페이스 소속 팀원 제거 API",
				"tags": [
					"[워크 스페이스 팀원] WorkspaceMember"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteWorkspaceMemberRequest"
							}
						}
					}
				}
			}
		},
		"/workspace-members/invitation": {
			"post": {
				"operationId": "InviteWorkspaceMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse"
								}
							}
						}
					}
				},
				"description": "** 0.2버전으로 미뤄짐 **\r\nSetting > 팀원 및 권한 > 팀 멤버 추가하기 (초대 메일 발송)",
				"tags": [
					"[세팅] Setting",
					"[워크 스페이스 팀원] WorkspaceMember"
				],
				"deprecated": true,
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InviteWorkspaceMemberRequest"
							}
						}
					}
				}
			}
		},
		"/workspace-members/invitation/{invitationId}": {
			"put": {
				"operationId": "UpdateInvitation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_WorkspaceInvitation_"
								}
							}
						}
					}
				},
				"description": "* ** 0.2버전으로 미뤄짐 **\r\n초대 메일 확인 후 승락/거절 처리 API",
				"tags": [
					"[세팅] Setting",
					"[워크 스페이스 팀원] WorkspaceMember"
				],
				"deprecated": true,
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "invitationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateInvitationRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}